fix: Wide ranging refactor to start "deslopping" this app
This project was precious to me and I let LLMs take the wheel far too much. This starts the crawl back

+++ internal/storage/postgres/strggen/models.go
@@ -32,5 +32,5 @@ type UserFoodlogentryMetric struct {
 	UserID         string
 	FoodlogentryID uuid.UUID
 	MetricKey      string
-	MetricValue    int32
+	MetricValue    float64
 }
